+++ examples/5. recover-from-stop/2. recover.ts
@@ -1,0 +1,5 @@
+import Spider from '../../src/core/Spider';
+import options from './options';
+
+
+new Spider(options).start();