+++ src/core/Downloader.ts
@@ -8,22 +8,30 @@ import Spider from './Spider';
 import { Agent as HttpAgent } from 'http';
 import { Agent as HttpsAgent } from 'https';
 import { Request } from 'node-fetch';
+import logger from '../util/logger';
 
 
 export default class Downloader {
 
     private _running: boolean = false
     private _idle: boolean = true
+    private _downloading: number = 0
+
+    get downloading() {
+        return this._downloading;
+    }
 
     constructor(
         private _spider: Spider,
         private _agents: Array<HttpAgent | HttpsAgent> = [],
-        private _handler: Handler) {
+        private _handler: Handler,
+        private _pipelines: Array<(items: Items) => void>) {
 
         _spider.on('canFetch', () => {
             this._idle = false;
             this._spider.active && this.start();
         });
+        _spider.on('complete', () => this._downloading--)
 
         _agents.push(_agents[0]);
         _agents[0] = undefined as any;
@@ -45,6 +53,7 @@ export default class Downloader {
                 this._idle = true;
                 continue;
             }
+            this._downloading++;
 
             this.request(uri);
         }
@@ -67,11 +76,22 @@ export default class Downloader {
         this._spider.emit('handle', url);
         let items = new Items(res);
         // distribute response
-        handler.handle.call(this._spider, res, items);
+        await handler.handle.call(this._spider, res, items);
+        this.pipe(items);
     }
     private currentAgent(useAgent?: boolean) {
         // Using pseudo random numbers to evenly access agents
         let current = _.random(1, this._agents.length - 1, false);
         return useAgent ? this._agents[current] : undefined;
+    }
+
+    async pipe(items: Items) {
+        if (this._pipelines) {
+            this._spider.emit('pipe', items.$response.url);
+            for (let line of this._pipelines) {
+                await line(items);
+            }
+        }
+        this._spider.emit('complete', items.$response.url);
     }
 }